WalletProject/requirements.txt
asgiref==3.4.1
backports.zoneinfo==0.2.1
certifi==2021.10.8
charset-normalizer==2.0.9
crypto==1.4.1
cryptoaddress==0.2.1
Django==2.2.2
djangorestframework==3.9.4
ecdsa==0.17.0
idna==3.3
Naked==0.1.31
Pillow==8.4.0
psycopg2==2.8.6
pycryptodome==3.12.0
pysha3==1.0.2
python-decouple==3.5
pytz==2019.1
PyYAML==6.0
requests==2.26.0
shellescape==3.8.1
six==1.16.0
sqlparse==0.3.0
urllib3==1.26.7
